212901/002/CODE ERROR FIXING
The issue detected by Codacy is related to the switch statement in the 'updateProduct.java' file.
The reason for the change is that switch statements should be exhaustive to make their control flow easier to follow. In this case, a default case or a switch branch for each possible value of the variable being switched on should be added.